+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ### 5.1.14
 
 + [x] Updated `Farbe` - fixed `saturate()`, `desaturate()`, and `brighten()`
++ [x] Fixed incorrect behavior of `calendar` and `calendarpicker` components, issue #1991